initial commit.
Basic files to create a wixsharp wpf that shows how the translations work.
It enforces the "es" language to be used. 
Shows that in a Custom Action there are two issues: the translation from standard existing tag is not translated to "es" and that new added tags are not recognized.